requirements.txt
dnsviz>=0.6.5
dnspython>=1.15.0
M2Crypto>=0.27.0
nagiosplugin>=1.2.4
pygraphviz
